refactor: fix useInput hook
